Update OptimizePage to use MVVM
